Code cleanup of message generation, managed to remove the import_statement hack

+++ messages/msg.h
@@ -1,10 +1,10 @@
 /*
-CAUV 2009/2010
-Author: Clemens Wiltsche
+    CAUV 2009/2010
+    Author: Clemens Wiltsche
 */
 
 /*
-Syntax-tree definition for the message creation.
+   Syntax-tree definition for the message creation.
 */
 
 #include <stdint.h>
@@ -12,53 +12,53 @@ Syntax-tree definition for the message creation.
 
 class Node
 {
-	public:
-	void print();
+    public:
+        void print();
 };
 
 class Declaration : public Node
 {
-	private:
-	char *m_type;
-	char *m_name;
+    private:
+        char *m_type;
+        char *m_name;
 
-	public:
-	Declaration(char*, char*);
+    public:
+        Declaration(char*, char*);
 
-	char* getName();
-	char* getType();
-	void print();
+        char* getName();
+        char* getType();
+        void print();
 };
 
 class Message : public Node
 {
-	private:
-	uint32_t m_id;
-	char *m_name;
-	std::vector<Declaration*>* m_declarations;
+    private:
+        uint32_t m_id;
+        char *m_name;
+        std::vector<Declaration*>* m_declarations;
 
-	public:
-	Message(uint32_t, char*, std::vector<Declaration*>*);
+    public:
+        Message(uint32_t, char*, std::vector<Declaration*>*);
 
-	uint32_t getId();
+        uint32_t getId();
         char* getName();
-        Declaration* getDeclaration(int);
-	void print();
+        Declaration* getDeclaration(size_t);
+        void print();
 
 };
 
 class Group : public Node
 {
-	private:
-	uint32_t m_id;
-	char *m_name;
-	std::vector<Message*>* m_messages;	
+    private:
+        uint32_t m_id;
+        char *m_name;
+        std::vector<Message*>* m_messages;	
 
-	public:
-	Group(uint32_t, char*, std::vector<Message*>*);
+    public:
+        Group(uint32_t, char*, std::vector<Message*>*);
 
-	uint32_t getId();
-	char* getName();
-	Message* getMessage(int);
-	void print();
+        uint32_t getId();
+        char* getName();
+        Message* getMessage(size_t);
+        void print();
 };
